test/requirements.txt
backcall==0.1.0
behave==1.2.6
boto3==1.9.96
botocore==1.12.96
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
docutils==0.14
ecdsa==0.13
envs==1.3
future==0.17.1
idna==2.8
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jmespath==0.9.3
parse==1.11.1
parse-type==0.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pycryptodome==3.3.1
Pygments==2.3.1
PyHamcrest==1.9.0
python-dateutil==2.8.0
python-jose-cryptodome==1.3.2
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
warrant==0.6.1
wcwidth==0.1.7
